Fix wrong time zone handling in telemetry
