+++ enhydris/telemetry/types/meteoview2.py
@@ -80,6 +80,7 @@ class TelemetryAPIClient(TelemetryAPIClientBase):
 
     def _get_start_date(self, sensor_id, timeseries_end_date):
         if timeseries_end_date is not None:
+            timeseries_end_date = timeseries_end_date.replace(tzinfo=None)
             start_date = timeseries_end_date + dt.timedelta(minutes=1)
         else:
             start_date = dt.datetime(1990, 1, 1)
